requirements.txt
pandas
PyYAML
Requests
snowflake_connector_python
pyarrow
sqlalchemy